unit-tests: hub-recycle test - move under hw-reset, use devices.enable_only, fail on missing hub/port
