+++ unit-tests/live/hw-reset/pytest-hub-recycle-imu.py
@@ -1,0 +1,109 @@
+# License: Apache 2.0. See LICENSE file in root directory.
+# Copyright(c) 2026 RealSense, Inc. All Rights Reserved.
+
+# Hub-recycle stress test: power-cycles the USB port the device under test is
+# connected to (true cold plug-in via the OS, not an in-band hardware_reset),
+# then verifies that the re-enumerated device exposes its full sensor set,
+# in particular the Motion Module on IMU-bearing devices.
+#
+# This catches the partial-enumeration race where, on Windows, the device-
+# watcher's debounce fires a "device added" callback before the HID Sensor
+# Collection has bound at T0+~1s, surfacing a UVC-only device with no IMU and
+# triggering "No HID info provided, IMU is disabled" / "HID Motion Sensor
+# Failure! bad optional access" in the log.
+#
+# Uses the existing rspy.devices.enable_only(recycle=True) flow so the hub
+# singleton and per-device port resolution are reused from the test fixtures.
+# Fails loudly (does not skip) if the hub or port cannot be resolved - this
+# test only makes sense on a CI agent with a controllable hub attached.
+
+import pytest
+import pyrealsense2 as rs
+from rspy import devices
+import logging
+log = logging.getLogger(__name__)
+
+
+pytestmark = [
+    pytest.mark.device_each("D435i"),
+    pytest.mark.device_each("D455"),
+    pytest.mark.device_each("D436"),
+    pytest.mark.device_each("D430i"),
+    pytest.mark.context("nightly"),
+    pytest.mark.timeout(3600),
+]
+
+
+ITERATIONS_NIGHTLY = 20
+ITERATIONS_WEEKLY  = 100
+
+
+def _sensor_names( dev ):
+    return [ s.get_info( rs.camera_info.name ) for s in dev.query_sensors() ]
+
+
+def _find_device_by_sn( ctx, sn ):
+    for d in ctx.query_devices():
+        if d.supports( rs.camera_info.serial_number ) and d.get_info( rs.camera_info.serial_number ) == sn:
+            return d
+    return None
+
+
+def test_hub_recycle_imu_presence( test_device, test_context_var ):
+    dev, ctx = test_device
+    sn = dev.get_info( rs.camera_info.serial_number )
+
+    # devices.enable_only(recycle=True) silently falls back to hw_reset when
+    # there's no hub - that would test something different from what this
+    # test claims. Refuse to run unless a real hub is present.
+    if devices.hub is None:
+        pytest.fail( "Test requires a controllable USB hub (Acroname/Ykush/UniFi) - no hub detected." )
+    if devices.get( sn ).port is None:
+        pytest.fail( f"Hub is present but could not resolve a port for serial {sn} - "
+                     "refusing to recycle all ports to avoid disturbing other devices on the hub." )
+
+    iterations = ITERATIONS_WEEKLY if 'weekly' in test_context_var else ITERATIONS_NIGHTLY
+    log.info( f"Hub-recycle IMU-presence test: {iterations} iterations on serial {sn}, "
+              f"hub port {devices.get( sn ).port}" )
+
+    missing_imu = []
+    no_reappear = []
+
+    for i in range( 1, iterations + 1 ):
+        log.debug( f"[{i}/{iterations}] recycling port" )
+        try:
+            # Recycles just this device's port (disable -> wait_until_removed
+            # -> enable -> wait_for re-enumeration), so other devices on the
+            # same hub are not disturbed.
+            devices.enable_only( [sn], recycle=True, timeout=devices.MAX_ENUMERATION_TIME )
+        except Exception as e:
+            log.error( f"[{i}/{iterations}] enable_only failed: {e}" )
+            no_reappear.append( i )
+            continue
+
+        new_dev = _find_device_by_sn( ctx, sn )
+        if new_dev is None:
+            log.error( f"[{i}/{iterations}] device with serial {sn} not in context after recycle" )
+            no_reappear.append( i )
+            continue
+
+        try:
+            names = _sensor_names( new_dev )
+        except RuntimeError as e:
+            log.error( f"[{i}/{iterations}] query_sensors raised: {e}" )
+            missing_imu.append( i )
+            continue
+
+        if any( "Motion" in n for n in names ):
+            log.debug( f"[{i}/{iterations}] OK: {names}" )
+        else:
+            log.error( f"[{i}/{iterations}] Motion Module MISSING: {names}" )
+            missing_imu.append( i )
+
+    if no_reappear:
+        log.error( f"Iterations that did not re-enumerate: {no_reappear}" )
+    if missing_imu:
+        log.error( f"Iterations with missing Motion Module: {missing_imu}" )
+
+    assert not no_reappear, f"{len(no_reappear)}/{iterations} iterations: device did not re-enumerate"
+    assert not missing_imu, f"{len(missing_imu)}/{iterations} iterations: device re-enumerated without Motion Module"
